implement _config.localization to avoid runtime import in ccalendar
