+++ pandas/_config/__init__.py
@@ -1,0 +1,4 @@
+"""
+pandas._config is considered explicitly upstream of everything else in pandas,
+should have no intra-pandas dependencies.
+"""
