+++ pandas/util/testing.py
@@ -18,6 +18,8 @@ import warnings
 import numpy as np
 from numpy.random import rand, randn
 
+from pandas._config.localization import set_locale
+
 from pandas._libs import testing as _testing
 import pandas.compat as compat
 from pandas.compat import (
@@ -527,38 +529,6 @@ def get_locales(prefix=None, normalize=True,
     pattern = re.compile('{prefix}.*'.format(prefix=prefix))
     found = pattern.findall('\n'.join(out_locales))
     return _valid_locales(found, normalize)
-
-
-@contextmanager
-def set_locale(new_locale, lc_var=locale.LC_ALL):
-    """Context manager for temporarily setting a locale.
-
-    Parameters
-    ----------
-    new_locale : str or tuple
-        A string of the form <language_country>.<encoding>. For example to set
-        the current locale to US English with a UTF8 encoding, you would pass
-        "en_US.UTF-8".
-    lc_var : int, default `locale.LC_ALL`
-        The category of the locale being set.
-
-    Notes
-    -----
-    This is useful when you want to run a particular block of code under a
-    particular locale, without globally setting the locale. This probably isn't
-    thread-safe.
-    """
-    current_locale = locale.getlocale()
-
-    try:
-        locale.setlocale(lc_var, new_locale)
-        normalized_locale = locale.getlocale()
-        if com._all_not_none(*normalized_locale):
-            yield '.'.join(normalized_locale)
-        else:
-            yield new_locale
-    finally:
-        locale.setlocale(lc_var, current_locale)
 
 
 def can_set_locale(lc, lc_var=locale.LC_ALL):
